Added controller, service, serviceImpl, mapper, request and response DTO of doctor

+++ backend/src/main/java/com/healthcare/backend/service/impl/DoctorServiceImpl.java
@@ -1,7 +1,12 @@
 package com.healthcare.backend.service.impl;
 
+import com.healthcare.backend.dto.request.DoctorRequestDTO;
+import com.healthcare.backend.dto.response.DoctorResponseDTO;
+import com.healthcare.backend.entity.Department;
 import com.healthcare.backend.entity.Doctor;
 import com.healthcare.backend.entity.User;
+import com.healthcare.backend.mapper.DoctorMapper;
+import com.healthcare.backend.repository.DepartmentRepository;
 import com.healthcare.backend.repository.DoctorRepository;
 import com.healthcare.backend.repository.UserRepository;
 import com.healthcare.backend.service.DoctorService;
@@ -20,35 +25,79 @@ public class DoctorServiceImpl implements DoctorService {
     @Autowired
     private UserRepository userRepository;
 
+    @Autowired
+    private DepartmentRepository departmentRepository;
+
+
     @Override
-    public Doctor addDoctor(Doctor doctor, Long userId) {
+    public DoctorResponseDTO addDoctor(
+            DoctorRequestDTO dto) {
 
-        User user = userRepository.findById(userId)
+        Doctor doctor =
+                DoctorMapper.toEntity(dto);
+
+
+        // Find User
+        User user = userRepository
+                .findById(dto.getUserId())
                 .orElseThrow(() ->
-                        new RuntimeException("User not found"));
+                        new RuntimeException(
+                                "User not found"));
 
-        if (!user.getRole().name().equals("DOCTOR")) {
-            throw new RuntimeException(
-                    "User must have DOCTOR role");
-        }
+
+        // Find Department
+        Department department =
+                departmentRepository
+                        .findById(dto.getDepartmentId())
+                        .orElseThrow(() ->
+                                new RuntimeException(
+                                        "Department not found"));
+
 
         doctor.setUser(user);
+        doctor.setDepartment(department);
 
-        return doctorRepository.save(doctor);
+
+        Doctor savedDoctor =
+                doctorRepository.save(doctor);
+
+        return DoctorMapper.toResponseDTO(
+                savedDoctor);
     }
+
 
     @Override
-    public List<Doctor> getAllDoctors() {
-        return doctorRepository.findAll();
+    public List<DoctorResponseDTO> getAllDoctors() {
+
+        return doctorRepository.findAll()
+                .stream()
+                .map(DoctorMapper::toResponseDTO)
+                .toList();
     }
 
+
     @Override
-    public Doctor getDoctorById(Long id) {
-        return doctorRepository.findById(id).orElse(null);
+    public DoctorResponseDTO getDoctorById(Long id) {
+
+        Doctor doctor =
+                doctorRepository.findById(id)
+                        .orElseThrow(() ->
+                                new RuntimeException(
+                                        "Doctor not found"));
+
+        return DoctorMapper.toResponseDTO(doctor);
     }
+
 
     @Override
     public void deleteDoctor(Long id) {
+
+        if (!doctorRepository.existsById(id)) {
+
+            throw new RuntimeException(
+                    "Doctor not found");
+        }
+
         doctorRepository.deleteById(id);
     }
 }
